ProjectManagerPlugin: fix compilation
Added missining file and reduced warnings.

+++ src/plugins/ProjectManager/ProjectManagerPlg.cpp
@@ -728,7 +728,7 @@ void ProjectManagerPlugin::do_runTask(const TaskInfo *task) {
 
     auto manager = getManager();
     auto count = manager->visibleTabs();
-    for (auto i = 0; i < count; i++) {
+    for (auto i = size_t(0); i < count; i++) {
         auto client = manager->getMdiClient(i);
         if (auto editor = dynamic_cast<qmdiEditor *>(client)) {
             editor->removeMetaData();
